Move primop in their own module
